refactor: decouple deployer widget into @nsite/deployer package
Extract the deploy/manage/auth component tree from apps/spa into a standalone
@nsite/deployer package so it can be consumed as an embeddable widget. The SPA
becomes a thin shell that mounts DeployerWidget and renders Navbar with
session props received via events.

Package layout:
- packages/deployer/src/components: 12 Svelte components (DeployerWidget,
  ManageSite, SuccessPanel, LoginModal, NIP46Dialog, AdvancedConfig,
  DeployZone, FileTree, ActivityRings, ProgressIndicator, LogoutConfirmModal,
  OperationBanner)
- packages/deployer/src/lib: upload, publish, nostr, scanner, crypto, files,
  base36, store (converted to createDeployerStores factory)
- packages/deployer/src/widget: NsiteDeployerElement HTMLElement wrapper and
  CSS string constant for the custom element
- packages/deployer/{vite.widget.config.js,vitest.config.js,svelte.config.js}

Workspace:
- Root package.json with npm workspaces (apps/spa, packages/deployer)
- .npmrc for @jsr registry
- deno.json nodeModulesDir: "none" for deno+npm workspace compat
- apps/spa/package.json depends on @nsite/deployer via workspace protocol
- CI installs deno workspace before build

Main fixes preserved by manual port into decoupled files (bypassing rebase
conflicts from PR #18 squash-merge divergence):
- DeployerWidget.svelte: merge kind-10063 blossoms into serverConfig.extraBlossoms
  so AdvancedConfig shows them; dTag regex allows interior hyphens; deploy
  blossom URL union no longer re-adds userBlossoms (already in extraBlossoms)
- packages/deployer/src/lib/upload.js: uploadBlobs and deleteBlobs batch auth
  signing (50 hashes per event) per BUD-11
- packages/deployer/src/lib/files.js: inferMimeType uses @std/media-types
  contentType() instead of inline MIME_TYPES map
