+++ .npmrc
@@ -1,0 +1,1 @@
+@jsr:registry=https://npm.jsr.io
