+++ deno.json
@@ -5,6 +5,7 @@
     "./apps/gateway",
     "./packages/shared"
   ],
+  "nodeModulesDir": "none",
   "imports": {
     "@std/assert": "jsr:@std/assert@^1.0.19"
   },
