+++ package.json
@@ -1,0 +1,7 @@
+{
+  "private": true,
+  "workspaces": [
+    "apps/spa",
+    "packages/deployer"
+  ]
+}
