+++ packages/deployer/src/components/ActivityRings.svelte
