+++ packages/deployer/src/components/FileTree.svelte
