+++ packages/deployer/src/components/LoginModal.svelte
@@ -1,7 +1,8 @@
 <script>
-  import { createEventDispatcher } from 'svelte';
-  import { createExtensionSigner, fetchProfile, DEFAULT_RELAYS } from '../lib/nostr.js';
-  import { session } from '../lib/store.js';
+  import { createEventDispatcher, getContext } from 'svelte';
+  import { createExtensionSigner, fetchProfile, DEFAULT_RELAYS } from '@nsite/deployer/nostr';
+
+  const { session } = getContext('deployer-stores');
   import { npubEncode } from 'nostr-tools/nip19';
   import NIP46Dialog from './NIP46Dialog.svelte';
