+++ packages/deployer/src/components/NIP46Dialog.svelte
@@ -1,8 +1,9 @@
 <script>
-  import { onMount, tick, createEventDispatcher } from 'svelte';
+  import { onMount, tick, createEventDispatcher, getContext } from 'svelte';
   import QRCode from 'qrcode';
-  import { createNostrConnectSigner, connectFromBunkerURI, fetchProfile, DEFAULT_RELAYS } from '../lib/nostr.js';
-  import { session } from '../lib/store.js';
+  import { createNostrConnectSigner, connectFromBunkerURI, fetchProfile, DEFAULT_RELAYS } from '@nsite/deployer/nostr';
+
+  const { session } = getContext('deployer-stores');
   import { npubEncode } from 'nostr-tools/nip19';
 
   export let show = false;
