+++ packages/deployer/src/components/ProgressIndicator.svelte
