+++ packages/deployer/src/lib/__tests__/scanner.test.js
