+++ packages/deployer/src/lib/nostr.js
