+++ packages/deployer/src/lib/scanner.js
